pyproject.toml
[tool.poetry]
name = "python-stream"
version = "0.1.0"
description = ""
authors = ["Carlos Kidman <carlos@qap.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
pyleniumio = "^1.13.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
